migrate js to ts
