--- tests/.setup/test-setup.js
@@ -1,35 +1,0 @@
-const base = require('@playwright/test');
-const { waitForRouteReady } = require('../../utils/waits/admin-page-ready');
-
-let page;
-
-exports.test = base.test.extend({
-  page: async ({ page }, use) => {
-    const originalGoto = page.goto.bind(page);
-    page.goto = async (url, options = {}) => {
-      const result = await originalGoto(url, {
-        waitUntil: 'domcontentloaded',
-        ...options,
-      });
-      await waitForRouteReady(page, url, { timeout: options.timeout || 15000 });
-      return result;
-    };
-
-    page.waitForRouteReady = async (url, options = {}) => {
-      await waitForRouteReady(page, url, options);
-    };
-
-    page.on('console', (msg) => {
-      const errorPattern = /error|failed|uncaught|exception|typeerror|referenceerror|syntaxerror|rangeerror|evalerror|urlerror|is not defined|cannot read|undefined|null is not an object/i;
-
-      if (msg.type() === 'error' || errorPattern.test(msg.text())) {
-        console.log(`[${msg.type().toUpperCase()}] ${msg.text()}`);
-      }
-    });
-
-    await use(page);
-  },
-});
-
-exports.expect = base.expect;
-exports.page = page;
