--- tests/mocks/ci/always-pass.mock.js
@@ -1,5 +1,0 @@
-const { test, expect } = require('@playwright/test');
-
-test('ci mock always passes', async () => {
-  expect(true).toBe(true);
-});
